tests/krate/show: Move tests into `routes` submodules
