+++ src/tests/krate/mod.rs
@@ -3,7 +3,6 @@ mod owners;
 mod publish;
 mod reverse_dependencies;
 mod search;
-mod show;
 mod summary;
 mod versions;
 mod yanking;
